start implementing a custom Externals
The MockExternal is unfortunately to rudimentary for what is necessary
out of the debugger here. In particular one thing is that we want to be
able to manage `Storage` separately from the lifetime of externals, and
another thing is that its `generate_data_id` for example is too
simplistic and does not necessarily reflect as closely what one would
expect from a real system.