store: add a mechanism to download and load store between runs
Convenient. Maybe a good idea would be to also maintain a log of them
and allow to go back to previous version with a button click? I really
don't wanna implement a full trie though...

But if we could build the near_store for wasm and use the trie with a
custom backend???????